+++ src/converter.ts
@@ -1,7 +1,9 @@
-import { Marpit, MarpitOptions } from '@marp-team/marpit'
+import { Marp, MarpOptions } from '@marp-team/marp-core'
+import { MarpitOptions } from '@marp-team/marpit'
 import * as chromeFinder from 'chrome-launcher/dist/chrome-finder'
 import puppeteer, { PDFOptions } from 'puppeteer-core'
 import { warn } from './cli'
+import { Engine } from './engine'
 import { error } from './error'
 import { File, FileType } from './file'
 import templates, { TemplateResult } from './templates/'
@@ -13,8 +15,8 @@ export enum ConvertType {
 
 export interface ConverterOption {
   allowLocalFiles: boolean
-  engine: typeof Marpit
-  html?: boolean
+  engine: Engine
+  html?: MarpOptions['html']
   lang: string
   options: MarpitOptions
   output?: string
@@ -128,7 +130,7 @@ export class Converter {
     const opts: any = { ...options, ...mergeOptions }
 
     // for marp-core
-    if (html !== undefined) opts.html = !!html
+    if (html !== undefined) opts.html = html
 
     const engine = new this.options.engine(opts)
 
@@ -136,7 +138,7 @@ export class Converter {
       error('Specified engine has not implemented render() method.')
 
     // for Marpit engine
-    engine.markdown.set({ html: !!html })
+    if (!(engine instanceof Marp)) engine.markdown.set({ html: !!html })
 
     return engine
   }